Create computer_difficult_path.py
There are two main components to this: A* pathfinding and max pooling. A* pathfinding looks for the "path of least resistance" from a starting point to an ending point. Max pooling (currently in main) creates a smaller field that is of a lower resolution, so that the A* algorithm can be more efficient. Individually, both portions have some functionality (not tested super closely yet), but I haven't had the chance to test them on top of one another.